Re-organize JS Folder and Reducers (#521)
* re-organize the js folder

* refactor reducers
