+++ app/src/js/__test__/redux.test.js
@@ -1,6 +1,6 @@
 /* @flow */
-import configureStore from '../functional/store/configure_store';
-import * as types from '../functional/actions/action_types';
+import configureStore from '../store/configure_store';
+import * as types from '../actions/action_types';
 import {ActionCreators} from 'redux-undo';
 
 let testJson = {
