+++ app/src/js/__test__/state.test.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import {makeLabel} from '../states';
+import {makeLabel} from '../functional/states';
 import _ from 'lodash/fp';
 
 describe('Label Tests', function() {
