+++ app/src/js/actions/action_creators.js
@@ -1,6 +1,6 @@
 /* @flow */
 import * as types from './action_types';
-import type {ItemType, LabelType} from '../../types';
+import type {ItemType, LabelType} from '../functional/types';
 
 /**
  * Create Item from url with provided creator
