+++ app/src/js/viewers/image_viewer.js
