+++ app/src/js/viewers/title_bar_viewer.js
